.hf-space/caption-studio-api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.7.2
ffmpeg-python==0.2.0
python-dotenv==1.0.1
aiofiles==23.2.1
boto3
